feat: make the webpack env validator a package
