+++ .gitignore
@@ -6,3 +6,4 @@ node_modules
 .turbo
 .husky/
 .vercel
+/packages/verify-env/dist/
