+++ packages/builder/package.json
@@ -76,7 +76,8 @@
     "url": "^0.11.0",
     "wagmi": "^0.6.4",
     "web-vitals": "^2.1.4",
-    "yup": "^0.32.11"
+    "yup": "^0.32.11",
+    "verify-env": "*"
   },
   "scripts": {
     "start": "craco start",
