+++ packages/verify-env/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "verify-env",
+  "version": "1.0.0",
+  "description": "",
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "test": "vitest --run",
+    "build": "pkgroll"
+  },
+  "keywords": [],
+  "author": "Atris",
+  "license": "ISC",
+  "devDependencies": {
+    "@tsconfig/node-lts": "^18.12.5",
+    "@types/node": "^20.5.9",
+    "@types/webpack": "^5.28.2",
+    "pkgroll": "^1.11.0",
+    "typescript": "^5.2.2",
+    "vitest": "^0.34.2"
+  }
+}
