+++ packages/verify-env/src/index.test.ts
@@ -1,0 +1,1 @@
+import { it, expect } from "vitest";
